requirements.txt
streamlit==1.31.1
pandas==2.1.4
numpy==1.24.3
plotly==5.18.0
scipy==1.11.4
python-docx==1.0.1
openpyxl==3.1.2